# backend/openapi.yml
openapi: 3.1.0
info:
  title: Fahrtenbuch API
  version: 1.0.0-dev
  description: |
    REST-API für das selbst gehostete Fahrtenbuch. Die API umfasst
    Authentifizierung, MFA, Passkeys, App-Pairing, Geräteverwaltung,
    Benutzer, Fahrzeuge, Fahrten, Tracking, Statistiken, Import/Export,
    Geocoding sowie zentrale Kartendienst-Konfiguration.
  license:
    name: GNU Affero General Public License v3.0
    identifier: AGPL-3.0-only
servers:
  - url: /api/v1
    description: Relativ zum Fahrtenbuch-Server
security:
  - bearerAuth: []
...
paths:
...
  /settings:
    get:
      tags: [Settings]
      summary: Eigene Einstellungen abrufen
      responses:
        '200': { description: Einstellungen }
    put:
      tags: [Settings]
      summary: Eigene Einstellungen aktualisieren
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200': { description: Einstellungen aktualisiert }
...
  /admin/users:
    get:
      tags: [Admin]
      summary: Benutzer auflisten
      responses:
        '200':
          description: Benutzer
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
    post:
      tags: [Admin]
      summary: Benutzer anlegen
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminUserInput' }
      responses:
        '201':
          description: Benutzer angelegt
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
  /admin/users/{id}:
    get:
      tags: [Admin]
      summary: Benutzer abrufen
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Benutzer
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
    put:
      tags: [Admin]
      summary: Benutzer aktualisieren
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200': { description: Benutzer aktualisiert }
    delete:
      tags: [Admin]
      summary: Benutzer löschen
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204': { description: Benutzer gelöscht }
...
  /admin/settings/services/map:
    put:
      tags: [Admin]
      summary: Kartenquelle konfigurieren
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MapServiceSettings' }
      responses:
        '200': { description: Kartenquelle gespeichert }
  /admin/settings/services/photon:
    put:
      tags: [Admin]
      summary: Photon konfigurieren
      description: Interne Server-URLs bleiben ausschließlich im Backend gespeichert.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PhotonServiceSettings' }
      responses:
        '200': { description: Photon-Konfiguration gespeichert }
  /admin/settings/services/overpass:
    put:
      tags: [Admin]
      summary: Overpass konfigurieren
      description: Interne Server-URLs bleiben ausschließlich im Backend gespeichert.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OverpassServiceSettings' }
      responses:
        '200': { description: Overpass-Konfiguration gespeichert }
  /admin/settings/services/test:
    post:
      tags: [Admin]
      summary: Kartendienst-Verbindung testen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category, config]
              properties:
                category: { type: string, enum: [map, photon, overpass] }
                config: { type: object, additionalProperties: true }
      responses:
        '200':
          description: Testergebnis
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success: { type: boolean }
                  responseTimeMs: { type: [integer, 'null'] }
                  message: { type: string }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    User:
      type: object
      required: [id, email, role, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        loginName: { type: [string, 'null'] }
        displayName: { type: [string, 'null'] }
        role: { type: string, enum: [user, admin] }
        locale: { type: string, default: de }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    AdminUserInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 8 }
        displayName: { type: [string, 'null'] }
        role: { type: string, enum: [user, admin], default: user }
...
    PublicMapConfig:
      type: object
      required: [provider, type, attribution, minZoom, maxZoom, tileSize]
      properties:
        provider: { type: string, enum: [osm, maplibre, custom] }
        type: { type: string, enum: [raster, vector, style] }
        tileUrl: { type: [string, 'null'], format: uri }
        styleUrl: { type: [string, 'null'], format: uri }
        attribution: { type: string }
        minZoom: { type: integer }
        maxZoom: { type: integer }
        tileSize: { type: integer, enum: [256, 512] }
    MapServiceSettings:
      allOf:
        - $ref: '#/components/schemas/PublicMapConfig'
    PhotonServiceSettings:
      type: object
      required: [provider, timeoutMs]
      properties:
        provider: { type: string, enum: [public, custom] }
        baseUrl: { type: [string, 'null'] }
        timeoutMs: { type: integer, minimum: 1000, default: 10000 }
    OverpassServiceSettings:
      type: object
      required: [provider, timeoutMs, searchRadiusMeters, maxResults]
      properties:
        provider: { type: string, enum: [public, custom] }
        interpreterUrl: { type: [string, 'null'] }
        timeoutMs: { type: integer, minimum: 1000, default: 30000 }
        searchRadiusMeters: { type: integer, minimum: 1, default: 2500 }
        maxResults: { type: integer, minimum: 1, maximum: 500, default: 50 }
...